fix media files uploading
